Cleaned up server and client files to get them ready for cross compilation. Basically I moved the functions into their own files (lin_client and lin_server.hh) and set them with namespaces. Now I'm going to see if I can get identical files to lin_server and lin_client with windows instead of linux.
